Fix interrupting image alignment in script
Also removed for now the custom Lua callback support to `imppg.align_images`. The way it was implemented was causing silent memory corruption (until trying to stop the script via Lua's `error`). The callback was running on a different thread than the Lua interpreter; granted, the interpreter was stopped at that time, but it's a wrong way to do it. Investigate lua_lock/unlock, lua_newthread.